.history/requirements_20230421180737.txt
numpy==1.19.5
scikit-learn==0.24.2
gensim==4.0.1
nltk==3.6.2
spaCy==3.1.2
Flask==2.0.2
Flask-CORS
transformers
torch
weaviate-client
python-dotenv